feat: Line의 points 부분 전략패턴으로 변경
