+++ src/main/java/ladder/domain/PointsStrategy.java
@@ -1,0 +1,7 @@
+package ladder.domain;
+
+import java.util.List;
+
+public interface PointsStrategy {
+    List<Boolean> points();
+}
