+++ src/test/java/ladder/LineTest.java
@@ -1,8 +1,12 @@
 package ladder;
 
 import ladder.domain.Line;
+import ladder.domain.RandomPoints;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,8 +15,7 @@ public class LineTest {
     @DisplayName("플레이어 수에 맞게 한 라인의 포인트 생성")
     @Test
     void 플레이어수_라인생성() {
-        int countOfPlayers = 4;
-        Line line = new Line(countOfPlayers);
+        Line line = createRandomLine(4);
         assertThat(line.numberOfPoints()).isEqualTo(4);
     }
 
@@ -20,15 +23,20 @@ public class LineTest {
     @DisplayName("라인의 마지막 점은 항상 false다")
     @Test
     void 라인_마지막점_false() {
-        int countOfPlayers = 4;
-        Line line = new Line(countOfPlayers);
-        assertThat(line.hasRightLine(countOfPlayers - 1)).isFalse();
+        Line line = createRandomLine(4);
+        assertThat(line.hasRightLine(3)).isFalse();
     }
 
     @DisplayName("point가 true면 오른쪽 가로 라인이 존재한다.")
     @Test
     void 라인_가로라인여부() {
-        Line line = new Line(true, false, true, false);
+        List<Boolean> points = Arrays.asList(true, false, true, false);
+        Line line = new Line(() -> points);
         assertThat(line.hasRightLine(0)).isTrue();
     }
+
+    private static Line createRandomLine(int countOfPlayers) {
+        return new Line(new RandomPoints(countOfPlayers));
+    }
+
 }
